# openapi.yaml
openapi: "3.0.2"
info:
  title: Todos API
  description: Documentation of a simple Todos API
  version: 0.0.1
servers:
  - url: https://soryo-fastify-todo.herokuapp.com
    description: Production
  - url: localhost
    description: Development server
tags:
  - name: todo
    description: Todo related end-points
  - name: tag
    description: Tag related end-points
  - name: healthcheck
    description: Tag for internal healthcheck routes

paths:
  /v1/todos:
    get:
      summary: List all the todos
      description: List all the todos of the user provided in the query parameters
      operationId: listTodos
      tags:
        - todo
      parameters:
        - name: ownerUuid
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successfully retrieved all the todos
          content:
            "application/json":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/todo"
              example:
                - uuid: 3d780d09-c520-4817-b430-ce849bcc5423
                  ownerUuid: 535d6711-2ec0-4ba7-9f34-3d13f25de822
                  title: Groceries
                  state: ACTIVE
    post:
      summary: Create a new todo
      description: Create a new todo
      operationId: createTodo
      tags:
        - todo
      requestBody:
        description: New todo payload
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/todo"
      responses:
        "200":
          description: Successfully created a new todo
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/todo"

  /v1/todos/{todoUuid}:
    get:
      summary: Get a todo
      description: Get a todo by providing its uuid
      operationId: getTodo
      tags:
        - todo
      parameters:
        - name: ownerUuid
          in: query
          required: true
          schema:
            type: string
        - name: todoUuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Successfully retrieved the todo
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/todo"
        "404":
          description: Couldn't find the todo with the provided uuid
    put:
      summary: Update a todo
      description: Update a todo by providing its uuid and the updated todo content
      operationId: updateTodo
      tags:
        - todo
      parameters:
        - name: ownerUuid
          in: query
          required: true
          schema:
            type: string
        - name: todoUuid
          required: true
          in: path
          schema:
            type: string
      requestBody:
        description: New todo payload
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/todo"
      responses:
        "200":
          description: Successful response
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/todo"
    delete:
      summary: Delete a todo
      description: Delete a todo by providing its uuid
      operationId: deleteTodo
      tags:
        - todo
      parameters:
        - name: ownerUuid
          in: query
          required: true
          schema:
            type: string
        - name: todoUuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/todo"

  /ready:
    get:
      summary: Get the readiness of the service
      operationId: isReady
      tags:
        - healthcheck
      responses:
        "200":
          description: Service is ready
        "503":
          description: Service isn't ready yet
  /healthy:
    get:
      summary: Get the healthiness of the service
      operationId: isHealthy
      tags:
        - healthcheck
      responses:
        "200":
          description: Service is healthy
        "503":
          description: Service isn't healthy
components:
  schemas:
    todo:
      type: object
      properties:
        uuid:
          description: The unique identifier of this todo
          type: string
        ownerUuid:
          description: The unique identifier of the owner of this todo
          type: string
        title:
          description: The title/short summary of this todo
          type: string
        state:
          description: The current state of this todo
          type: string
          enum:
            - "ACTIVE"
            - "COMPLETED"
            - "DELETED"
        description:
          description: The lengthy description of this todo
          type: string
      required:
        - uuid
        - "ownerUuid"
        - "title"
        - "state"
      example:
        uuid: 3d780d09-c520-4817-b430-ce849bcc5423
        ownerUuid: 535d6711-2ec0-4ba7-9f34-3d13f25de822
        title: Groceries
        state: ACTIVE